+++ core/datastore/src/main/java/com/threegap/bitnagil/datastore/auth/crypto/TokenCrypto.kt
@@ -1,0 +1,6 @@
+package com.threegap.bitnagil.datastore.auth.crypto
+
+interface TokenCrypto {
+    fun encrypt(bytes: ByteArray): ByteArray
+    fun decrypt(bytes: ByteArray): ByteArray
+}
